+++ lib/generators/puffer/puffer_user/templates/puffer_user.rb
@@ -1,0 +1,3 @@
+# class PufferUser < Puffer::User::Mongoid
+class PufferUser < Puffer::User::ActiveRecord
+end