+++ lib/generators/puffer/puffer_user/templates/puffer_users_controller.rb
@@ -1,3 +1,2 @@
 class Admin::PufferUsersController < Puffer::PufferUsersBase
-  unloadable
 end