+++ config/routes.rb
@@ -1,7 +1,5 @@
-Rails.application.routes.draw do
-
+Puffer::Engine.routes.draw do
   namespace :admin do
     resource :session, :only => [:new, :create, :destroy]
   end
-
 end