+++ lib/puffer/backends/controllers/sessions/clearance.rb
@@ -21,7 +21,7 @@ class Puffer::Sessions::Clearance < Puffer::Sessions::Base
 
   def destroy
     sign_out
-    redirect_to new_admin_session_url
+    redirect_to puffer.new_admin_session_url
   end
 
 private