+++ lib/puffer/engine.rb
@@ -1,5 +1,7 @@
 module Puffer
   class Engine < Rails::Engine
+    engine_name :puffer
+
     config.autoload_paths << File.join(root, 'lib')
 
     initializer 'puffer_components.set_configs', :after => 'action_controller.set_configs' do |app|